Refactor and clean dataset preprocessing funcs
